fix image menu hack #44
