+++ src/menu/menu-image.js
@@ -1,0 +1,43 @@
+import {elt} from 'prosemirror/dist/util/dom'
+import {MenuItem} from 'prosemirror/dist/menu/menu'
+import {focusedIndex, isCollapsed} from '../util/pm'
+
+
+function run (pm) {
+  const index = focusedIndex(pm)
+  if (index == null) return
+  pm.ed.trigger('command.menu.file', index)
+}
+
+function select (pm) {
+  return isCollapsed(pm)
+}
+
+function render (pm) {
+  const el = elt('div'
+  , { style:
+      { cursor: 'pointer' }
+    }
+  , 'Upload Image'
+  )
+  el.addEventListener('mousedown', function (event) {
+    // HACK around #44
+    event.stopPropagation()
+    event.stopImmediatePropagation()
+  })
+  el.addEventListener('click', function (event) {
+    run(pm)
+  })
+  return el
+}
+
+const menuImage = new MenuItem(
+  { label: 'Upload Image'
+  , title: 'upload image(s) above this block'
+  , run
+  , select
+  , render
+  }
+)
+
+export default menuImage
